requirements.txt
fastapi==0.111.0
mindee==4.6.0
openai==1.33.0
pydantic==2.7.3
python-dotenv==1.0.1
bcrypt==4.1.3
python_jose==3.3.0
pytz==2024.1
resend==2.0.0
SQLAlchemy==2.0.25
